// rs-css/src/interface/enums.ts
enum Classes {
    CONTAINER = 'container',
    WRAPPER = 'wrapper',
    LEVEL = 'level',
    PASSED = 'passed',
    NOT_PASSED = 'not-passed',
    LEVEL_HEADER = 'level-header',
    LEVEL_TITLE = 'level-title',
    LIST_ACTIVE = 'list-active',
    LEVEL_HELP = 'level-help',
    LEVEL_MENU = 'level-menu',
    LEVEL_LIST = 'level-list',
    LIST_ITEM = 'list-item',
    SELECTOR_NAME = 'selector-name',
    TITLE = 'title',
    SYNTAX = 'syntax',
    DESCRIPTION = 'description',
    CHECK_MARK = 'check-mark',
    CHECK = 'check',
    BURGER = 'burger',
    OPEN = 'open',
    EXAMPLES = 'examples',
    EXAMPLE = 'example',
    RESET = 'reset',
    GAME = 'game',
    EDITOR = 'editor',
    EDITOR_WRAPPER = 'editor-wrapper',
    EDITOR_MAIN = 'editor-main',
    VIEWER = 'viewer',
    VIEWER_MAIN = 'viewer-main',
    NUMBERS = 'numbers',
    HTML_CODE = 'html-code',
    LAYOUT = 'layout',
    LAYOUT_HEADER = 'layout-header',
    HEADING_WRAPPER = 'heading-wrapper',
    TABLE_PLANETS = 'table-planets',
    TOOLTIP = 'tooltip',
    FORM = 'form',
    FORM_BUTTON = 'form-button',
    FORM_INPUT = 'form-input',
    FORM_HELP = 'form-help',
    BLINK = 'blink',
    SHAKE = 'shake',
    SELECTED = 'selected',
    BUTTON_TEXT = 'button-text',
    TABLE_WRAPPER = 'table-wrapper',
    TABLE = 'table',
    BUTTON_MENU = 'button-menu',
    BUTTON_PREV = 'button-prev',
    BUTTON_NEXT = 'button-next',
    RS_SCHOOL = 'rs-school',
    YEARS = 'years',
    GIT = 'git',
    FOOTER = 'footer',
    BACKLIGHT = 'backlight',
    ACTIVE = 'active',
    HIDDEN = 'hidden',
    SCORE = 'score',
    GAME_WIN = 'game-win',
    CORRECT = 'correct',
    INCORRECT = 'incorrect',
    MISSED = 'missed',
    TITLE_RESULTS = 'results',
}

enum Tags {
    H2 = 'h2',
    H3 = 'h3',
    DIV = 'div',
    SPAN = 'span',
    P = 'p',
    BUTTON = 'button',
    UL = 'ul',
    LI = 'li',
    FORM = 'form',
    I = 'i',
    INPUT = 'input',
    SECTION = 'section',
    FOOTER = 'footer',
}

enum Text {
    CSS = 'CSS Editor',
    STYLE = 'style.css',
    HTML = 'HTML viewer',
    INDEX = 'index.html',
    EXAMPLES = 'Examples',
    RESET = 'RESET',
    PLACEHOLDER = 'Type in a CSS selector',
    ENTER = 'ENTER',
    HELP = 'HELP',
    GITHUB = 'github',
    LINK_GITHUB = 'https://github.com/alexpashchuk',
    RS = '© RS-CSS 2023',
    LINK_RS = 'https://rs.school/js/',
    YOU_WIN = 'YOU WIN',
    CORRECT = 'correct',
    INCORRECT = 'incorrect',
    MISSED = 'missed',
    RESULTS = 'Results:',
}

export { Classes, Tags, Text };